requirements.txt
openai-whisper
pyannote.audio
python-dotenv
torch
torchaudio
